[FEAT] 회원-스토어 팔로우 조회, 취소 API 수정

+++ src/main/java/com/lotte/danuri/member/follow/FollowService.java
@@ -9,7 +9,7 @@ import javax.persistence.FetchType;
 
 public interface FollowService {
 
-    int register(Long memberId, Long storeId);
+    Long register(Long memberId, Long storeId);
 
     List<StoreDto> getStoresByFollow(Long memberId);
